Add files via upload
I created a Timed Robot code template for our Programming Lectures with team, to use it during the lectures as the main material. Here are properties of my code:
1. We have two autonomous options. At the defaul one robot is basically moving straight for five meters. 
And the second option makes robot move straight for five meters, turn left 90 degrees and keep moving until it gets close to some obstacle. 
2. For teleop we use logitech joysticks. 
Button X: activates intake motor as long as we press the button. (1 VictorSPX, 70% speed)
Button Y: activates shooter motors as long as we press the button. (2 VictorSPX'es, 80% speed)
Button A: Once we press the button, elevator motors run for 5 seconds. With the second press they run in opposite direction for 5 seconds again. (2 VictorSPX'es, 80% speed)
Button B: Once we press the button, solenoid pushes the cover of the intake and keeps pushing it, keeps it open. With the second press solenoid turns off.